+++ examples/SPI/ReadGconf/ReadGconf.ino
@@ -13,7 +13,7 @@ SPIClass & spi = SPI;
 const uint8_t CHIP_SELECT_PIN = 10;
 
 const long SERIAL_BAUD_RATE = 115200;
-const int DELAY = 1000;
+const uint16_t DELAY = 1000;
 
 // Instantiate TMC51X0
 TMC51X0 tmc5160;
