+++ examples/SPI/TuneStealthChop/TuneStealthChop.ino
@@ -15,7 +15,7 @@ const uint8_t ENABLE_HARDWARE_PIN = 4;
 
 const long SERIAL_BAUD_RATE = 115200;
 const int LOOP_DELAY = 500;
-const int DELAY = 4000;
+const uint16_t DELAY = 4000;
 
 // converter constants
 // internal clock is ~12MHz
